Create a separate container for the Hearts component on the search page
