+++ web/src/js/components/Search/SearchMenuContainer.js
@@ -6,14 +6,14 @@ import SearchMenu from 'js/components/Search/SearchMenuComponent'
 export default createFragmentContainer(SearchMenu, {
   app: graphql`
     fragment SearchMenuContainer_app on App {
-      ...HeartsContainer_app
+      ...SearchHeartsContainer_app
       ...MoneyRaisedContainer_app
     }
   `,
   user: graphql`
     fragment SearchMenuContainer_user on User {
       id
-      ...HeartsContainer_user
+      ...SearchHeartsContainer_user
     }
   `,
 })
