+++ web/src/js/constants.js
@@ -2,7 +2,6 @@
   Configuration
 **/
 
-export const MAX_DAILY_HEARTS_FROM_SEARCHES = 150
 export const MAX_DAILY_HEARTS_FROM_TABS = 150
 
 /**
